KTOR-4236 Add CORS logging
